requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.1
Django==3.2.10
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.6
mimesis==4.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
zipp==3.6.0
